requirements.txt
numpy==1.19.1
Pillow==7.2.0
PySocks==1.7.1
requests==2.24.0
requests-oauthlib==1.3.0
tweepy==3.9.0
urllib3==1.25.10
